Dutch translation
I saw a few other translations so I thought I add the Dutch translation as well, is easier for the end user to have it all in one package. We still have it in our repository at: https://github.com/dutch-translators I will add a note in the readme file with a note and link that the translations is added to the core after this pr is merged.